style: move general styles to another location
