+++ src/styles/pages/general.js
@@ -1,9 +1,10 @@
 import styled from 'styled-components';
 import media from 'styled-media-query';
-import { themes } from './themes';
-import SocialMedias from '../components/social-medias';
 import AniLink from 'gatsby-plugin-transition-link/AniLink';
 import Img from 'gatsby-image';
+
+import { themes } from '../themes';
+import SocialMedias from '../../components/social-medias';
 
 export const PagesWrap = styled.div`
   margin: auto;
